+++ CHANGELOG.md
@@ -9,7 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Features
 
+- program: add reduce only user status ([#560](https://github.com/drift-labs/protocol-v2/pull/560))
+
 ### Fixes
+
+- program: add update_lp_market_position test for big k ([#565](https://github.com/drift-labs/protocol-v2/pull/565))
 
 ### Breaking